fix(jni): wrong integer type conversion

+++ app/src/main/java/com/osfans/trime/ui/fragments/KeyboardFragment.kt
@@ -49,7 +49,7 @@ class KeyboardFragment :
                 Trime.getServiceOrNull()?.recreateInputView()
             }
             "keyboard__candidate_page_size" -> {
-                val pageSize = AppPrefs.defaultInstance().keyboard.candidatePageSize.toInt()
+                val pageSize = AppPrefs.defaultInstance().keyboard.candidatePageSize
                 if (pageSize <= 0) return
                 lifecycleScope.launch {
                     withContext(Dispatchers.IO) {
